requirements.txt
requests
lxml
python-dotenv
google-generativeai